Adding file mode. Formatting fix
In this commit, I added the first version of the program in file mode (arguments are not read from the console but from the file).  A file with arguments args.txt was also added in which a, b and c are still being read from a new line (perhaps this version of the file will remain). Source code formatted with go fmt.

The code that waits for "q" to exit has been added for future reference when the program is compiled and run on the command line.

The problem with error handling in interactive mode is still relevant. Also I need to implement error handling when a wrong file is entered.